changing the project file environment
