+++ src/components/Task6/style.css
@@ -1,0 +1,3 @@
+.cange-fonts-italic {
+    font-style: italic;
+}
