+++ src/components/Task7/style.css
@@ -1,0 +1,3 @@
+.login-check {
+    background-color: cadetblue;
+}
